+++ app/Enums/ActivityTypes.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Enums;
+
+enum ActivityTypes: string
+{
+    case COOLIFY_PROCESS = 'coolify_process';
+}
