+++ tests/Unit/RulesTest.php
@@ -1,0 +1,5 @@
+<?php
+
+test('globals')
+    ->expect(['dd', 'dump', 'ray'])
+    ->not->toBeUsed();
